tbcc/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
telethon>=1.34.0
python-telegram-bot>=21.0
celery[redis]>=5.3.0
redis>=5.0.0
httpx>=0.26.0
alembic>=1.13.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
